minor code cleanup
