--- src/tasks/IProgress.ts
@@ -1,3 +1,0 @@
-export interface IProgress {
-  report(progress: number): void;
-}
