--- src/tasks/UploadResponseHandlerOption.ts
@@ -1,6 +1,0 @@
-import { ResponseHandler, ResponseHandlerOption } from "@microsoft/kiota-abstractions";
-import { UploadResponseHandler } from "./UploadResponseHandler";
-
-export class UploadResponseHandlerOption extends ResponseHandlerOption {
-  public responseHandler?: ResponseHandler = new UploadResponseHandler();
-}
