--- src/tasks/UploadSession.ts
@@ -1,6 +1,0 @@
-export interface UploadSession {
-  expirationDateTime?: Date | null;
-  nextExpectedRanges?: string[] | null;
-  odataType?: string | null;
-  uploadUrl?: string | null;
-}
